(GH-108) Changed rule from Guideline to Requirment
